+++ src/Sentry.ts
@@ -1,0 +1,16 @@
+import * as Sentry from '@sentry/browser'
+import { Integrations } from '@sentry/tracing'
+import tenants from './tenants'
+
+export default {
+  init: (formsAppId: number, options: Sentry.BrowserOptions) => {
+    Sentry.init({
+      integrations: [new Integrations.BrowserTracing()],
+      ...options,
+    })
+    Sentry.setTag('tenant', tenants.tenant)
+    Sentry.setTag('formsAppId', formsAppId)
+  },
+  captureException: Sentry.captureException,
+  setTag: Sentry.setTag,
+}