+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `Sentry` to allow for error capturing
+
 ## [0.7.2] - 2020-04-15
 
 ### Added